Improved DetailsActivty Layout, Fixed Image Scaling
Everything is near-complete. Scaling issue with the images loaded on the
DetailsActivity was fixed... mostly. No testing has been done on screens
smaller than 1920 x 1080, needs to be done. The layout of that activity
itself has also been improved, and it is the 2nd thing remaining to be
done (other thing being adding a refresh button to the action bar).

Handling lack of connectivity is still not fully done, but there is no
more time at this point.